FEAT (Navbar) : adds Mobile version of the Nav
